+++ src/components/ControlsContent/ControlsContent.tsx
@@ -1,11 +1,13 @@
 import Button from "components/common/Button";
 import React from "react";
-import Settings from "components/common/Settings";
+import Settings from "components/Settings";
 import WinsLossesButton from "../WinsLossesButton";
 import { useData } from "context";
 import useStyles from "./ControlsContent.styles";
+import { useTranslation } from "react-i18next";
 
 export default function ControlsContent(): JSX.Element {
+  const { t } = useTranslation();
   const classes = useStyles();
   const [showSettings, setShowSettings] = React.useState(false);
   const {
@@ -36,12 +38,12 @@ export default function ControlsContent(): JSX.Element {
       <div className={classes.scoreButtons}>
         <div className={classes.titleContainer}>
           <h4 className={classes.title} />
-          <h4 className={classes.title}>Wins</h4>
-          <h4 className={classes.title}>Losses</h4>
+          <h4 className={classes.title}>{t("controls.wins")}</h4>
+          <h4 className={classes.title}>{t("controls.losses")}</h4>
         </div>
 
         <div className={classes.scoreButtonsSection}>
-          <h4>Innocent</h4>
+          <h4>{t("controls.innocent")}</h4>
           <WinsLossesButton
             buttonBackgroundColor={theme.innocentTextColor}
             decrement={() => setInnocentWins(innocentWins - 1)}
@@ -59,7 +61,7 @@ export default function ControlsContent(): JSX.Element {
         </div>
 
         <div className={classes.scoreButtonsSection}>
-          <h4>Impostor</h4>
+          <h4>{t("controls.impostor")}</h4>
           <WinsLossesButton
             buttonBackgroundColor={theme.impostorTextColor}
             decrement={() => setImpostorWins(impostorWins - 1)}
@@ -79,13 +81,13 @@ export default function ControlsContent(): JSX.Element {
 
       <div className={classes.buttonContainer}>
         <Button classNames={classes.reset} onClick={() => resetScores()}>
-          Reset Scores
+          {t("controls.resetScores")}
         </Button>
         <Button
           classNames={classes.reset}
           onClick={() => resetPlayersPositions()}
         >
-          Reset Round
+          {t("controls.resetRound")}
         </Button>
       </div>
       <div className={classes.buttonContainer}>
@@ -93,10 +95,10 @@ export default function ControlsContent(): JSX.Element {
           classNames={`${classes.reset} ${classes.dangerButton}`}
           onClick={() => resetAll()}
         >
-          Reset All
+          {t("controls.resetAll")}
         </Button>
       </div>
-      <h2>Notes</h2>
+      <h2>{t("controls.notes")}</h2>
       <div className={classes.notesContainer}>
         <textarea
           className={classes.notes}
@@ -107,13 +109,13 @@ export default function ControlsContent(): JSX.Element {
         />
       </div>
       <Button classNames={`${classes.resetNotes}`} onClick={() => setNotes("")}>
-        Reset Notes
+        {t("controls.resetNotes")}
       </Button>
       <Button
         classNames={`${classes.dangerButton}`}
         onClick={() => setShowSettings(true)}
       >
-        Settings
+        {t("controls.settings")}
       </Button>
 
       <Settings show={showSettings} onClose={() => setShowSettings(false)} />