+++ src/components/PlayersSection/PlayersSection.tsx
@@ -4,6 +4,7 @@ import Section from "components/Section";
 import { useData } from "context";
 import useStyles from "./PlayersSection.styles";
 import { MobileContext } from "components/App";
+import { useTranslation } from "react-i18next";
 
 export default function PlayersSection(): JSX.Element {
   const {
@@ -19,6 +20,7 @@ export default function PlayersSection(): JSX.Element {
     setUnknownPlayers,
   } = useData()!; // eslint-disable-line
 
+  const { t } = useTranslation();
   const classes = useStyles();
   const isMobile = React.useContext(MobileContext);
 
@@ -30,40 +32,31 @@ export default function PlayersSection(): JSX.Element {
 
   const sections: Array<Section> = [
     {
-      title: "Innocent",
+      title: t("main.lists.innocent"),
       list: innocentPlayers,
       setList: setInnocentPlayers,
     },
     {
-      title: "Suspicious",
+      title: t("main.lists.suspicious"),
       list: susPlayers,
       setList: setSusPlayers,
     },
     {
-      title: "Evil / Hit List",
+      title: `${t("main.lists.evil")} / ${t("main.lists.hitList")}`,
       list: evilPlayers,
       setList: setEvilPlayers,
     },
     {
-      title: "Dead",
+      title: t("main.lists.dead"),
       list: deadPlayers,
       setList: setDeadPlayers,
     },
     {
-      title: "Unknown",
+      title: t("main.lists.unknown"),
       list: unknownPlayers,
       setList: setUnknownPlayers,
     },
   ];
-
-  // function swapPlayersColors(
-  //   currentPlayerColor: string,
-  //   targetPlayerColor: string,
-  //   currentPlayerList: Array<IPlayer>
-  // ) {
-  //   // const currentPlayer = currentPlayerList.filter()
-  //   console.log(currentPlayerList);
-  // }
 
   return (
     <div className={classes.root}>