+++ src/components/Recovery/Recovery.styles.ts
@@ -1,6 +1,8 @@
+import { ITheme } from "utils/types";
 import { createUseStyles } from "react-jss";
+import { hexToRGB } from "utils/colorConverter";
 
-export default createUseStyles({
+export default createUseStyles((theme: ITheme) => ({
   footer: {
     display: "flex",
     alignItems: "end",
@@ -12,4 +14,14 @@ export default createUseStyles({
   left: {
     marginLeft: "0.5rem",
   },
-});
+  dangerButton: {
+    backgroundColor: theme.buttonDangerBackgroundColor,
+    color: theme.buttonDangerTextColor,
+
+    "&:hover": {
+      backgroundColor: `rgba(${hexToRGB(
+        theme.buttonDangerBackgroundColor
+      )}, 0.75)`,
+    },
+  },
+}));